refactor: files separated in directories by types
