+++ Frontend/src/styles/TaskManager.css
